feat: move write infra file into gcp package
