+++ cli/cmd/bootstrap_gcp.go
@@ -4,7 +4,6 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -149,18 +148,14 @@ func (c *BootstrapGcpCmd) BootstrapGcp() error {
 		c.CodesphereEnv.FeatureFlags[flag] = true
 	}
 
-	bootstrapErr := bs.Bootstrap()
+	err = bs.Bootstrap()
+	writeGcpInfraFile(bs)
 
-	err = writeInfraDetails(bs.Env)
 	if err != nil {
-		log.Printf("warning: failed to write infra details: %v", err)
-	}
-
-	if bootstrapErr != nil {
 		if bs.Env.Jumpbox != nil && bs.Env.Jumpbox.GetExternalIP() != "" {
 			log.Printf("To debug on the jumpbox host:\nssh-add $SSH_KEY_PATH; ssh -o StrictHostKeyChecking=no -o ForwardAgent=yes -o SendEnv=OMS_PORTAL_API_KEY root@%s", bs.Env.Jumpbox.GetExternalIP())
 		}
-		return fmt.Errorf("failed to bootstrap GCP: %w", bootstrapErr)
+		return fmt.Errorf("failed to bootstrap GCP: %w", err)
 	}
 
 	log.Println("\n🎉🎉🎉 GCP infrastructure bootstrapped successfully!")
@@ -181,28 +176,11 @@ func (c *BootstrapGcpCmd) BootstrapGcp() error {
 	return nil
 }
 
-// writeInfraDetails writes details about the bootstrapped codesphere environment into a file.
-func writeInfraDetails(csEnv *gcp.CodesphereEnvironment) error {
-	envBytes, err := json.MarshalIndent(csEnv, "", "  ")
+func writeGcpInfraFile(bs *gcp.GCPBootstrapper) {
+	err := bs.WriteInfraFile()
 	if err != nil {
-		return fmt.Errorf("failed to marshal codesphere env: %w", err)
+		log.Printf("warning: failed to write infra details: %v", err)
+	} else {
+		log.Printf("infrastructure details written to file: %s", gcp.GetInfraFilePath())
 	}
-
-	workdir := env.NewEnv().GetOmsWorkdir()
-	fw := util.NewFilesystemWriter()
-
-	err = fw.MkdirAll(workdir, 0755)
-	if err != nil {
-		return fmt.Errorf("failed to create workdir %w", err)
-	}
-
-	infraFilePath := gcp.GetInfraFilePath()
-	err = fw.WriteFile(infraFilePath, envBytes, 0644)
-	if err != nil {
-		return fmt.Errorf("failed to write gcp bootstrap env file: %w", err)
-	}
-
-	log.Printf("Infrastructure details written to %s", infraFilePath)
-
-	return nil
 }
