+++ internal/bootstrap/gcp/gcp.go
@@ -5,7 +5,6 @@ package gcp
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"slices"
@@ -170,31 +169,6 @@ func NewGCPBootstrapper(
 	}, nil
 }
 
-func GetInfraFilePath() string {
-	workdir := env.NewEnv().GetOmsWorkdir()
-	return fmt.Sprintf("%s/gcp-infra.json", workdir)
-}
-
-// LoadInfraFile reads and parses the GCP infrastructure file from the specified path.
-// Returns the environment, whether the file exists, and any error.
-// If the file doesn't exist, returns an empty environment with exists=false and nil error.
-func LoadInfraFile(fw util.FileIO, infraFilePath string) (CodesphereEnvironment, bool, error) {
-	if !fw.Exists(infraFilePath) {
-		return CodesphereEnvironment{}, false, nil
-	}
-
-	content, err := fw.ReadFile(infraFilePath)
-	if err != nil {
-		return CodesphereEnvironment{}, true, fmt.Errorf("failed to read gcp infra file: %w", err)
-	}
-
-	var env CodesphereEnvironment
-	if err := json.Unmarshal(content, &env); err != nil {
-		return CodesphereEnvironment{}, true, fmt.Errorf("failed to unmarshal gcp infra file: %w", err)
-	}
-	return env, true, nil
-}
-
 func (b *GCPBootstrapper) Bootstrap() error {
 	err := b.stlog.Step("Validate input", b.ValidateInput)
 	if err != nil {
@@ -216,9 +190,9 @@ func (b *GCPBootstrapper) Bootstrap() error {
 		return fmt.Errorf("failed to ensure GCP project: %w", err)
 	}
 
-	err = b.stlog.Step("Save install config checkpoint locally", b.CreateInstallConfigCheckpoint)
+	err = b.stlog.Step("Write infrastructure file", b.WriteInfraFile)
 	if err != nil {
-		return fmt.Errorf("failed to create install config checkpoint: %w", err)
+		return fmt.Errorf("failed to write infrastructure file: %w", err)
 	}
 
 	err = b.stlog.Step("Ensure billing", b.EnsureBilling)
